Improve handling of full file paths
Before we were passing a pointer from the alpm file path to the hashset.
This meant that we couldn't modify the file path to include the root.
Now, we're dynamically allocating an array in the scope of main().

This seems a bit faster than trying to remove the root from each
path to search as well.